requirements.txt
chromadb>=0.5.0
pinecone>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
matplotlib>=3.7.0
streamlit>=1.35.0
